+++ back/src/main/java/com/tdevilleduc/urthehero/back/constant/ResilienceConstants.kt
@@ -3,6 +3,6 @@ package com.tdevilleduc.urthehero.back.constant
 object ResilienceConstants {
     const val INSTANCE_STORY_SERVICE = "story-service"
     const val INSTANCE_PAGE_SERVICE = "page-service"
-    const val INSTANCE_PERSON_SERVICE = "person-service"
+    const val INSTANCE_USER_SERVICE = "user-service"
     const val INSTANCE_PROGRESSION_SERVICE = "progression-service"
 }