+++ back/src/test/java/com/tdevilleduc/urthehero/back/util/TestUtil.kt
@@ -1,15 +1,12 @@
 package com.tdevilleduc.urthehero.back.util
 
-import com.tdevilleduc.urthehero.back.model.PageDTO
-import com.tdevilleduc.urthehero.back.model.PersonDTO
-import com.tdevilleduc.urthehero.back.model.Story
-import com.tdevilleduc.urthehero.back.model.StoryDTO
+import com.tdevilleduc.urthehero.back.model.*
 import org.testcontainers.shaded.org.apache.commons.lang.RandomStringUtils
 import java.util.*
 
 object TestUtil {
     private val random: Random = Random()
-    fun createRandomStory(): Story {
+    fun createStory(): Story {
         val storyAuthorId = random.nextInt()
         val storyFirstPageId = random.nextInt()
         val title = RandomStringUtils.random(20)
@@ -18,7 +15,7 @@ object TestUtil {
         return Story(0, storyAuthorId, storyFirstPageId, title, detailedText, image)
     }
 
-    fun createStory(authorId: Int, firstPageId: Int): StoryDTO {
+    fun createStoryDto(authorId: Int, firstPageId: Int): StoryDTO {
         val id = null
         val title = RandomStringUtils.random(20)
         val detailedText = RandomStringUtils.random(20)
@@ -26,19 +23,31 @@ object TestUtil {
         return StoryDTO(id, title, authorId, firstPageId, detailedText, image)
     }
 
-    fun createPerson(): PersonDTO {
+    fun createUserDto(): UserDTO {
         val id = null
-        val login = RandomStringUtils.random(20)
-        val displayName = RandomStringUtils.random(20)
-        val email = RandomStringUtils.random(20)
+        val username = RandomStringUtils.random(20)
         val password = RandomStringUtils.random(20)
-        return PersonDTO(id, login, displayName, email, password)
+        return UserDTO(id, username, password)
     }
 
-    fun createPage(): PageDTO {
+    fun createUser(): User {
+        val id = null
+        val username = RandomStringUtils.random(20)
+        val password = RandomStringUtils.random(20)
+        return User(id, username, password)
+    }
+
+    fun createPageDto(): PageDTO {
         val id = null
         val text = RandomStringUtils.random(20)
         val image = RandomStringUtils.random(20)
         return PageDTO(id, text, image)
+    }
+
+    fun createPage(): Page {
+        val id = 0
+        val text = RandomStringUtils.random(20)
+        val image = RandomStringUtils.random(20)
+        return Page(id, text, image)
     }
 }